Update Easy_Webscraping_Exercise.ipynb
Updated the extraction of `temp_state`.
Search based on the word `low` seems to be more stable. Previously (search based on `high`), `low` and `high` were detected in the same sentence, which resulted in a wrong output (a low being reported as high).
They could probably be a much better way to make the search more robust.